requirements.txt
geoalchemy2==0.4.2
sqlalchemy==1.2.6
pyyaml==3.12
